+++ configs/vitest.config.js
@@ -1,0 +1,23 @@
+import {
+    configDefaults,
+    coverageConfigDefaults,
+    defineConfig,
+} from "vitest/config";
+
+export default defineConfig({
+    test: {
+        globals: true,
+        exclude: [...configDefaults.exclude, "dist/**/*"],
+        coverage: {
+            provider: "v8",
+            exclude: [
+                ...coverageConfigDefaults.exclude,
+                "dist/**/*",
+                "src/types/*",
+                "src/configs/*",
+            ],
+            reporter: ["text", "html", "clover", "json", "lcov"],
+        },
+    },
+    plugins: [],
+});