+++ stylegen/vitest.config.ts
@@ -1,16 +1,3 @@
-import {
-    configDefaults,
-    coverageConfigDefaults,
-    defineConfig,
-} from "vitest/config";
+import configs from "@portfolio/configs/vitest";
 
-export default defineConfig({
-    test: {
-        exclude: [...configDefaults.exclude, "dist/**/*"],
-        coverage: {
-            provider: "v8",
-            exclude: [...coverageConfigDefaults.exclude, "dist/**/*"],
-        },
-    },
-    plugins: [],
-});
+export default configs;