+++ ui/eslint.config.js
@@ -1,0 +1,8 @@
+import createEslintConfig from "@portfolio/configs/eslint";
+
+export default createEslintConfig(
+    "tsConfig",
+    "tsVueConfig",
+    "tsTestConfig",
+    "prettierConfig",
+);